requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==3.4.8
elasticsearch==7.15.0
gidgethub==5.0.1
idna==3.2
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-dotenv==0.19.0
regex==2021.9.30
starlette==0.14.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
yapf==0.31.0
yarl==1.6.3
